[FE/style] Community_Screen : CREATE community tab - PostWidget (#47)
